backend/requirements.txt
# Web Framework
fastapi
uvicorn[standard]
python-multipart

# Async Tasks
celery
redis

# Database
sqlalchemy
psycopg2-binary
alembic

# Audio/Video Processing
faster-whisper
pyannote.audio
pydub
librosa
noisereduce
soundfile
ffmpeg-python
demucs  # Separação de voz e instrumentos para música

# NLP and AI
openai
deepmultilingualpunctuation
spacy

# Storage
minio
boto3

# Utilities
python-dotenv
pydantic
pydantic-settings
httpx
python-json-logger

# Monitoring
prometheus-client
sentry-sdk

# Testing
pytest
pytest-asyncio
pytest-cov

# Development
black
ruff
mypy
